test_requirements.txt
teamcity-messages
pytz
django-debug-toolbar